requirements.txt
# requirements.txt

tensorflow==2.15.0
ultralytics==8.3.119
opencv-python-headless==4.8.1.78
gradio
numpy==1.26.0
requests==2.31.0
scikit-image==0.21.0
Pillow==10.0.1
python-dotenv==1.0.0
selenium
folium

git+https://github.com/facebookresearch/detectron2.git

